separate parts

+++ processing/cache.py
@@ -1,13 +1,17 @@
 from io import BytesIO
 from math import floor, ceil
+from copy import copy
 
 from django.core.files.base import ContentFile
 import numpy as np
+
 from PIL import Image
 
-from .utils import Submatrix, exr_to_array, image_from_array
+from processing.models import BodySource
+from .utils import Matrix, Submatrix, exr_to_array, image_from_array
 
 RENDER_SIZE = (4096, 4096)
+
 
 class CacheBuilder(object):
     SCALE_MAP = {
@@ -18,8 +22,8 @@ class CacheBuilder(object):
     }
 
     EXR_FIELDS = ('uv',)
-    RGB_FIELDS = ('light', 'ao')
-    RGBA_FIELDS = ('body',)
+    RGBA_FIELDS = ('body', 'light', 'ao')
+    L_FIELDS = ('uv_alpha',)
 
     @staticmethod
     def create_cache(instance, fields, cache_model):
@@ -35,11 +39,13 @@ class CacheBuilder(object):
                 size = array.shape[:2]
                 array[..., 0] *= size[0]
                 array[..., 1] *= size[1]
-                array[..., 0] %= 1024
-                array[..., 1] %= 1024
-            submatrix = Submatrix(array)
+
+            if isinstance(instance.content_object, BodySource):
+                matrix = Matrix(array)
+            else:
+                matrix = Submatrix(array)
             scale = CacheBuilder.SCALE_MAP.get(field, 1)
-            matrices.append((field, submatrix, scale))
+            matrices.append((field, matrix, scale))
 
         size_array = []
         for _, matrix, scale in matrices:
@@ -49,26 +55,41 @@ class CacheBuilder(object):
 
         bbox = (min(x0), min(y0), max(x1), max(y1))
 
+        if 'uv' in fields:
+            (_, matrix, scale) = next(mx for mx in matrices if mx[0] == 'uv')
+            alpha = copy(matrix)
+            alpha.values = matrix.values[..., 3]
+            alpha._source = matrix._source[..., 3]
+            matrix.values = matrix.values[..., :2]  # cut redundant channels from matrix
+            matrices.append(('uv_alpha', alpha, scale))
+
         for field, matrix, scale in matrices:
             # remove old cache
             instance.cache.filter(source_field=field).delete()
             scaled_bbox = tuple(x * scale for x in bbox)
-            cache = cache_model(source_field=field, source=instance, position=scaled_bbox[:2])
             matrix.repick(scaled_bbox)
-            (buffer, extension) = CacheBuilder.get_buffer(field, matrix)
+
+            (buffer, extension) = CacheBuilder.get_buffer(field, matrix.values)
             filename = "%s_%s_%s.%s" % (instance._meta.model_name, instance.id, field, extension)
+            position = tuple(int(x) for x in scaled_bbox[:2])[::-1]
+            cache = cache_model(source_field=field, source=instance, position=position)
             cache.file.save(filename, ContentFile(buffer.getvalue()))
 
     @staticmethod
-    def get_buffer(field, matrix):
+    def get_buffer(field, array):
         buffer = BytesIO()
         if field in CacheBuilder.EXR_FIELDS:
             extension = 'npy'
-            np.save(buffer, matrix.values)  # TODO: only save [:2] slice and save alpha separately
+            np.save(buffer, array)
+        elif field in CacheBuilder.L_FIELDS:
+            extension = 'png'
+            array = (array * 255.0).astype('uint8')
+            img = Image.fromarray(array, mode='L')
+            img.save(buffer, extension)
         else:
             extension = 'png'
             channels = ('R', 'G', 'B', 'A') if field in CacheBuilder.RGBA_FIELDS else ('R', 'G', 'B')
-            img = image_from_array(matrix.values, channels=channels)
+            img = image_from_array(array, channels=channels)
             img.save(buffer, extension)
 
         return (buffer, extension)
@@ -80,7 +101,7 @@ class CacheBuilder(object):
 
         img = Image.open(texture.texture.path)
         size = img.size
-        tiled_size = (x/texture.tiling for x in RENDER_SIZE)
+        tiled_size = tuple(x / texture.tiling for x in RENDER_SIZE)
         if size != tiled_size:
             img = img.resize(tiled_size, Image.LANCZOS)
 
@@ -90,4 +111,3 @@ class CacheBuilder(object):
         buffer.flush()
         filename = "%s.npy" % texture.texture.name
         texture.cache.save(filename, ContentFile(buffer.getvalue()))
-
